+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Fixed bug with GFW deforestation alert clicking behavior
  - Improve documentation and add MIT license
  - Better UI for adding Planet imagery in sidepanel
- - Added 2023 + 2024 Planet imagery to viewer
+ - Added 2023 + 2024 Planet imagery to imagery viewer
+ - Added ability to download imagery
+ - Added enhanced analytics
 
 ## [0.1.1] - 2025-07-14